double carlories

+++ Data/Entities/User/Ingredient.cs
@@ -36,7 +36,7 @@ public class Ingredient
     [Display(Name = "Serving Size (grams)")]
     public int ServingSizeGrams { get; set; }
 
-    public int CaloriesPerServing { get; set; }
+    public double CaloriesPerServing { get; set; }
 
     public int GramsPerCup { get; set; }
