+++ Lib/ViewModels/Newsletter/IngredientViewModel.cs
@@ -25,7 +25,7 @@ public class IngredientViewModel
 
     public int GramsPerCup { get; init; }
 
-    public int CaloriesPerServing { get; set; }
+    public double CaloriesPerServing { get; set; }
 
     public Allergy Allergens { get; init; }
